Add single instance support on top of blitzdb
The blitzdb get() method by default return different class instances,
so that if two instances of the same document (same pk) are retrieved,
changes don't propagate - if you change an attribute of instance 2,
tha value of the attribute will still be the same for instance 1.
the change will show up only if instance 2 is saved to the database last
and then both instances reloaded from the database.

While this behavior is fine in many cases, there are other use cases
where Tsubame wants to get a single instances of a document,
because it could be retrieved from the database and modified in many places.
An example of this is the local user list - if there are separate instances of the
backing data and the user list is modified in one places, the change will
not reflect in other places (such as filters using the list) until Tsubame
is restarted. There is also the risk that the change will be overwritten
by what is basically stale data if any of the local user list instances
is saved to the database after the modified local user list instance is saved.

So we subclass the BlitzDB FileBackend and extend the get method by the
optional single_instance_parameter, which is set to False by default.
If the parameter is set to True, the get method will behave slightly differently:

- there is now a dictionary of single instances classes indexed by the data class pk
- first the parent class get() will be called
- pk of the result (if any) will be looked up in the dictionary
- if found, the instance from the dictionary will be returned
- if not found, the data class instance will be added to the dictionary and then returned

Like this it is possible to get Document instances that are single class program wide,
which should simplify a lot of thing and should mostly eliminate the need for separate instance
caching and or additional mechanisms for notification of backing data change.

There are a couple of potential issues to watch for though:

- if the same data class instance is retrieved by the get method with single_instance == False,
  it will be a different instance from the single instance one, so attribute changes will
  not propagate and the saved data state will depend on what instance saves last
- the filter function does not yet support single instance behavior