Add bundled copy of python-twitter and it's dependencies
This effectively adds a local copy of:

- python-twitter
- future
- requests
- requests_oauthlib

That's a rather big pile of code!
Unfortunately there is really a no way around that
on some (mostly mobile platforms), but it should
be possible to use system version of most of those
on more sane platforms, such as Fedora and other "normal" Linux distros.