Add persistence for message source classes
Turns out message source instances *do* need persistence
as any API usage requires the information about which account to use.

We might ass well add a persistent "enabled" property when we are at it,
which can be used to turn the source on an off. This can be useful for
temporarily turning a a source on an off, rather than adding and removing it from a stream.