Handle CLI tasks earlier
This is mainly to work around the platform detection messages ending in
the output, but it can also mean less unnecessary work is done before
Tsubame handles a task from CLI and shuts down.

If some CLI tasks need initialized platform module, we can split those
tasks into a "late" batch that runs after the platform module is
initialized. IIRC that's how this is done in modRana.