Add support for a separate Tweet cache db to the database manager
Looks like we will need some tweet caching for the Twitter message sources,
so as preparatory work add a new database instance to the db manager
for caching recent tweets.

This is done as Tweet caching could be rather high volume in comparison
to the main database that basically holds mostly general configuration
that on the other hand needs to be pretty robust and fast to access.

Also by having the twitter cache in a separate folder we can store it
in the designated XDG cache folder. The database is a cache so the system
can safely drop it and it will just be recreated from online sources.