Move CLI task handling after platform module instantiation
The CLI tasks might need the platform module (to be for example able to
open an URL in a platform independent way).

In case this turns out to be undesirable (something on CLI needs to run before platform module is instantiated or platform modules slows things down too much) we can always split the CLI task handling to two methods, one running before and one after platform module instantiation.