Always use fullscreen mode when viewing images
Drop the confusing and buggy header-visible/full-screen toggling and
just show images always in fullscreen.

The floating menu button with controls and the pull down menu with Silica make this possible. It's also still possible to get back directly to the previous page - there is the back button with controls and with silica the page stack indicator can be taped to do the same thing.