Use QtQuick Controls 2 as the Controls backend
Turns out Controls 2 provide numerous benefits to out use-case:
- much more sane scroll bar support
- better performance on mobile devices and overall
- looks to be generally cleaned up
- there seem to be quite a few new & possibly useful elements
At the same time it looks like we don't really use any of the stuff the removed/changed when moving from 1 to 2 (just some tweeks for the Menu element have been needed), so let's just switch and that's it. :)

Not sure about the Button style thing yet but buttons seem to show up normally even without it so we can likely leave this for later.