Add keyboard based scrolling
Yes, it needs to be this brain dead stupid
because QML ListView scrolling and focus
handling SUCKS.

We also do one pixel flicks as the return to bounds function
does not animate and as we actually can't detect an upcoming border
until we hit it. All in all a nice learning experience and
a pretty big exercise in frustration.

Also note the commit time (CEST)...