Add temporary message stream logic
This makes it possible to implement temporary message streams
for cases such as user clicking on a hashtag, user tweets, user favorites, etc.

For now hastags temporary streams are supported and more will be added.

Also add a shutdown signal that can be used by code residing in the Qt 5 GUI module to react to Tsubame shutdown.