Save pictures to monthly subfolders by default
If one saves many pictures from Twitter it can result with many pictures
that occupy a single folder and can be hard to browse, especially on
mobile devices. So save pictures to subfolders by default, where each subfolder corresponds to a month and is named in the <year>_<month> format.