Update bundled copy of python-twitter to 3.5
We were using 3.2.1 before so it's time to update.
3.5 should handle some API changes Twitter did in
in the last few months and I have not yet spotted
anything that is not working. Also this should
fix some cases where the 280 character tweets
have not been shown correctly.

And of course, distro packagers are free to drop the
bundled python-twitter copy and use a system-wide one. :)