Removed unused components originally from modRana
A lot of especially QML code in Tsubame actually comes
from the modRana GUI and there have still been some components
not really relevant for Tsubame or simply not yet used.

So let's remove those, even the ones we might need some day
as the component version in modRana might actually improve
in the meantime.