Add list membership processing functionality and improve caching
Make it possible to get membership-anotated listing of lists for a username, so that we can show on which lists a username already is and on which not. This can be used when implementating a add/remove from list view.

Also further improve the account user info caching code by moving the cache refresh code to a separate private method.