Add bundled copies of modules for Japanese processing
The furigana module can add furigana to kanji via the ruby markup
and jaconv is it's dependency.

Note that for this to work also mecab, mecab-iapdic and Python 3 bindings
for mecab need to be installed. All these three require compiled native
code, so can't be simply bundled but user needs to make sure they are
installed from appropriate sources.