Make it to show furigana to a twitter message
Furigana is basically the fonetic transcription of kanji,
which can help you read the kanji if you know the corresponding
japanese word but don't know what the given kanji stands for.

The feature works by adding a "show furigana" button when we
detect the given Twitter message is in Japanese. When user clicks the
button, the japanese message text is sent to Python, where the text is
processed by the furigana module which adds furigana to it via the
ruby markup. The result is returned as a simple html document
which we feed to a WebView, that overlaps everything on the MessagePage
other than the page and message header.

The furigana WebView can be hidden by cliking it, making the original message and other elements, such as the "show furigana" button, visisble again.