Add support for showing a plain text version of a Tweet
The Tweet page now has a new top-menu entry "Show as text" that opens the Tweet in plain text in a text editor. The text of the Tweet contains the Tweet URL, full Tweet plain text and links to any media.

This way it's easy to copy parts of the Tweet to clipboard or even to edit the Tweet and then copy it to somewhere else.

Resolves: #88 on GitHub