Add API for tracking if Page is on the application page stack
As unfrotunately both Silica & QtQuick Controls 2 page stack
can dependably garbage collect otherwise unrefferenced pages
we need to be able to track if a page is on stack and
possibly destroy it once it leaves the stack.

Otherwise the pages would simply accumulate as long as the
application is running, effectively leading to a memory leak
and would only be cleaned up once the app is closed
(or terminated by OOM killer :P).