Destroy pages on pop by default
We need to parent pages to rWin on creation or else they
could not be displayed (at least according to the QML docs),
but this has the apparent side effect that the page will
continue to exist even after it has been popped from the
page stack and Tsubame holds no direct reference on it.

This is effectively a memory leak and we need to do
something about is, especially as Tsubame can by design
pop many pages during normal usage scenarios as the user
goes through message streams, user info pages, Tweet details, etc.

So use the isOnPagestack property to detect that the page
has left the page stack and destroy it unless told otherwise.