Add the PlatformImagePicker element
Finalize the addition of the PlaformImagePicker by adding
a Silica variant and documenting the new element.

Also rename the property holding the selected file URLs to
selectedFiles, to no conflict with the platform specific
implementation, so that we can provide our own consistent one.

QtQuick.Dialogs returns an array while Silica.Pickers returns
a ListModel, so we have to make sure selectedFiles is always an array ourselves.