Work around rate limiting issues when fetching lists owned by a user
With the default configuration of python-twitter, it is possible to fetch about 300 lists per user before rate limits are hit, due to low (20) count used by the nested method.

A PR has been opened upstream to make the count configurable, in the meantime let's use a modified GetList() implementation with a high (500) count.

This way we should be able to avoid hitting rate limits even for users with an extremely high list count.